fix: push knowledge about the playlist down into the playlist component
