+++ frontend/src/scenes/session-recordings/playlist/Playlist.tsx
@@ -5,8 +5,19 @@ import { useActions, useValues } from 'kea'
 import { ReactNode, useRef, useState } from 'react'
 
 import { IconMagicWand } from '@posthog/icons'
-import { LemonButton, LemonCollapse, LemonSkeleton, LemonTag, Tooltip } from '@posthog/lemon-ui'
+import {
+    LemonBadge,
+    LemonBanner,
+    LemonButton,
+    LemonCollapse,
+    LemonSkeleton,
+    LemonTag,
+    Link,
+    Spinner,
+    Tooltip,
+} from '@posthog/lemon-ui'
 
+import { PropertyKeyInfo } from 'lib/components/PropertyKeyInfo'
 import { FEATURE_FLAGS } from 'lib/constants'
 import { useResizeBreakpoints } from 'lib/hooks/useResizeObserver'
 import { LemonTableLoader } from 'lib/lemon-ui/LemonTable/LemonTableLoader'
@@ -14,8 +25,15 @@ import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { range } from 'lib/utils'
 import { maxGlobalLogic } from 'scenes/max/maxGlobalLogic'
 import { DraggableToNotebook } from 'scenes/notebooks/AddToNotebook/DraggableToNotebook'
+import { useNotebookNode } from 'scenes/notebooks/Nodes/NotebookNodeContext'
+import { RecordingsUniversalFiltersEmbedButton } from 'scenes/session-recordings/filters/RecordingsUniversalFiltersEmbed'
+import { SessionRecordingPreview } from 'scenes/session-recordings/playlist/SessionRecordingPreview'
+import { SessionRecordingsPlaylistTopSettings } from 'scenes/session-recordings/playlist/SessionRecordingsPlaylistSettings'
+import { SessionRecordingsPlaylistTroubleshooting } from 'scenes/session-recordings/playlist/SessionRecordingsPlaylistTroubleshooting'
+import { sessionRecordingsPlaylistLogic } from 'scenes/session-recordings/playlist/sessionRecordingsPlaylistLogic'
+import { urls } from 'scenes/urls'
 
-import { SessionRecordingType } from '~/types'
+import { ReplayTabs, SessionRecordingType } from '~/types'
 
 const SCROLL_TRIGGER_OFFSET = 100
 
@@ -38,50 +56,24 @@ export type PlaylistContentBlock = PlaylistSectionBase & {
 export type PlaylistSection = PlaylistRecordingPreviewBlock | PlaylistContentBlock
 
 export type PlaylistProps = {
-    sections: PlaylistSection[]
-    listEmptyState: JSX.Element
     title?: string
-    notebooksHref?: string
-    embedded?: boolean
-    loading?: boolean
-    headerActions?: JSX.Element
-    footerActions?: JSX.Element
-    filterActions?: JSX.Element | null
-    onScrollListEdge?: (edge: 'top' | 'bottom') => void
+    type?: 'filters' | 'collection'
+    logicKey?: string
+    isSynthetic?: boolean
+    description?: string
     selectInitialItem?: boolean
-    onSelect?: (item: SessionRecordingType) => void
-    onChangeSections?: (activeKeys: string[]) => void
-    'data-attr'?: string
-    activeItemId?: string
 }
 
 export function Playlist({
     title,
-    notebooksHref,
-    loading,
-    embedded = false,
-    activeItemId: propsActiveItemId,
-    sections,
-    headerActions,
-    footerActions,
-    filterActions,
-    onScrollListEdge,
-    listEmptyState,
+    type,
+    logicKey,
+    isSynthetic,
+    description,
     selectInitialItem,
-    onSelect,
-    onChangeSections,
-    'data-attr': dataAttr,
 }: PlaylistProps): JSX.Element {
     const { featureFlags } = useValues(featureFlagLogic)
     const { askSidePanelMax } = useActions(maxGlobalLogic)
-
-    const firstRecordingItem = sections
-        .filter((s): s is PlaylistRecordingPreviewBlock => 'items' in s)
-        ?.find((s) => s.items.length > 0)?.items[0]
-
-    const [controlledActiveItemId, setControlledActiveItemId] = useState<SessionRecordingType['id'] | null>(
-        selectInitialItem && firstRecordingItem ? firstRecordingItem.id : null
-    )
 
     const playlistListRef = useRef<HTMLDivElement>(null)
     const { ref: playlistRef, size } = useResizeBreakpoints({
@@ -89,23 +81,32 @@ export function Playlist({
         750: 'medium',
     })
 
-    const onChangeActiveItem = (item: SessionRecordingType): void => {
-        setControlledActiveItemId(item.id)
-        onSelect?.(item)
-    }
-
-    const initiallyOpenSections = sections.filter((s) => s.initiallyOpen).map((s) => s.key)
-    const [openSections, setOpenSections] = useState<string[]>(initiallyOpenSections)
-
-    const onChangeOpenSections = (activeKeys: string[]): void => {
-        setOpenSections(activeKeys)
-        onChangeSections?.(activeKeys)
-    }
-
-    const activeItemId = propsActiveItemId === undefined ? controlledActiveItemId : propsActiveItemId
-
     const lastScrollPositionRef = useRef(0)
     const contentRef = useRef<HTMLDivElement | null>(null)
+
+    const notebookNode = useNotebookNode()
+    const embedded = !!notebookNode
+
+    // bound outside here
+    const {
+        filters,
+        activeSessionRecordingId,
+        totalFiltersCount,
+        sessionRecordingsResponseLoading,
+        pinnedRecordings,
+        otherRecordings,
+        hasNext,
+    } = useValues(sessionRecordingsPlaylistLogic)
+    const { maybeLoadSessionRecordings, setFilters, setSelectedRecordingId } =
+        useActions(sessionRecordingsPlaylistLogic)
+
+    const onScrollListEdge = (edge: 'bottom' | 'top'): void => {
+        if (edge === 'top') {
+            maybeLoadSessionRecordings('newer')
+        } else {
+            maybeLoadSessionRecordings('older')
+        }
+    }
 
     const handleScroll = (e: React.UIEvent<HTMLDivElement>): void => {
         if (e.currentTarget.scrollTop > lastScrollPositionRef.current) {
@@ -124,21 +125,99 @@ export function Playlist({
         lastScrollPositionRef.current = e.currentTarget.scrollTop
     }
 
+    const sections: PlaylistSection[] = []
+
+    if (type === 'collection' || pinnedRecordings.length > 0) {
+        sections.push({
+            key: 'pinned',
+            title: (
+                <div className="flex flex-row deprecated-space-x-1 items-center">
+                    <span>Pinned recordings</span>
+                    <LemonBadge.Number count={pinnedRecordings.length} status="muted" size="small" />
+                </div>
+            ),
+            items: pinnedRecordings,
+            render: ({ item, isActive }) => <SessionRecordingPreview recording={item} isActive={isActive} selectable />,
+            initiallyOpen: true,
+        })
+    } else {
+        sections.push({
+            key: 'other',
+            title: (
+                <div className="flex flex-row deprecated-space-x-1 items-center">
+                    <span>Results</span>
+                    <LemonBadge.Number count={otherRecordings.length} status="muted" size="small" />
+                </div>
+            ),
+            items: otherRecordings,
+            initiallyOpen: !pinnedRecordings.length,
+            render: ({ item, isActive }) => <SessionRecordingPreview recording={item} isActive={isActive} selectable />,
+            footer: (
+                <div className="p-4">
+                    <div className="h-10 flex items-center justify-center gap-2 text-secondary">
+                        {sessionRecordingsResponseLoading ? (
+                            <>
+                                <Spinner textColored /> Loading older recordings
+                            </>
+                        ) : hasNext ? (
+                            <LemonButton onClick={() => maybeLoadSessionRecordings('older')}>Load more</LemonButton>
+                        ) : (
+                            'No more results'
+                        )}
+                    </div>
+                </div>
+            ),
+        })
+    }
+
+    const firstItem = sections
+        .filter((s): s is PlaylistRecordingPreviewBlock => 'items' in s)
+        ?.find((s) => s.items.length > 0)?.items[0]
     const sectionCount = sections.length
     const itemsCount = sections
         .filter((s): s is PlaylistRecordingPreviewBlock => 'items' in s)
         .flatMap((s) => s.items).length
 
+    const initiallyOpenSections = sections.filter((s) => s.initiallyOpen).map((s) => s.key)
+    const [openSections, setOpenSections] = useState<string[]>(initiallyOpenSections)
+
+    const onChangeOpenSections = (activeKeys: string[]): void => {
+        setOpenSections(activeKeys)
+    }
+
+    const [controlledActiveItemId, setControlledActiveItemId] = useState<SessionRecordingType['id'] | null>(
+        selectInitialItem && firstItem ? firstItem.id : null
+    )
+
+    const onChangeActiveItem = (item: SessionRecordingType): void => {
+        setControlledActiveItemId(item.id)
+        setSelectedRecordingId(item.id)
+    }
+
+    const activeItemId = activeSessionRecordingId === undefined ? controlledActiveItemId : activeSessionRecordingId
+
+    const listEmptyState =
+        type === 'collection' ? (
+            <CollectionEmptyState isSynthetic={isSynthetic} description={description} />
+        ) : (
+            <ListEmptyState />
+        )
+
     return (
         <div className="flex flex-col min-w-60">
-            {filterActions && (
-                <DraggableToNotebook className="mb-2" href={notebooksHref}>
-                    {filterActions}
+            {!notebookNode && (
+                <DraggableToNotebook className="mb-2" href={urls.replay(ReplayTabs.Home, filters)}>
+                    <RecordingsUniversalFiltersEmbedButton
+                        filters={filters}
+                        setFilters={setFilters}
+                        totalFiltersCount={totalFiltersCount}
+                        currentSessionRecordingId={activeSessionRecordingId}
+                    />
                 </DraggableToNotebook>
             )}
             <div
                 ref={playlistRef}
-                data-attr={dataAttr}
+                data-attr="session-recordings-playlist"
                 className={clsx('Playlist w-full min-w-60 min-h-96', {
                     'Playlist--wide': size !== 'small',
                     'Playlist--embedded': embedded,
@@ -149,13 +228,18 @@ export function Playlist({
                     className="Playlist__list flex flex-col relative overflow-hidden h-full w-full"
                 >
                     <div className="flex flex-col relative w-full bg-bg-light overflow-hidden h-full Playlist__list">
-                        <DraggableToNotebook href={notebooksHref}>
+                        <DraggableToNotebook href={urls.replay(ReplayTabs.Home, filters)}>
                             <div className="flex flex-col gap-1">
                                 <div className="shrink-0 bg-bg-3000 relative flex justify-between items-center gap-0.5 whitespace-nowrap border-b">
                                     {title && <TitleWithCount title={title} count={itemsCount} />}
-                                    {headerActions}
+                                    <SessionRecordingsPlaylistTopSettings
+                                        filters={filters}
+                                        setFilters={setFilters}
+                                        type={type}
+                                        shortId={logicKey}
+                                    />
                                 </div>
-                                <LemonTableLoader loading={loading} />
+                                <LemonTableLoader loading={sessionRecordingsResponseLoading} />
                             </div>
                         </DraggableToNotebook>
                         <div className="overflow-y-auto flex-1" onScroll={handleScroll} ref={contentRef}>
@@ -169,7 +253,7 @@ export function Playlist({
                                             content: (
                                                 <SectionContent
                                                     section={s}
-                                                    loading={!!loading}
+                                                    loading={!!sessionRecordingsResponseLoading}
                                                     setActiveItemId={onChangeActiveItem}
                                                     activeItemId={activeItemId}
                                                     emptyState={listEmptyState}
@@ -186,19 +270,16 @@ export function Playlist({
                             ) : sectionCount === 1 ? (
                                 <SectionContent
                                     section={sections[0]}
-                                    loading={!!loading}
+                                    loading={!!sessionRecordingsResponseLoading}
                                     setActiveItemId={onChangeActiveItem}
                                     activeItemId={activeItemId}
                                     emptyState={listEmptyState}
                                 />
-                            ) : loading ? (
+                            ) : sessionRecordingsResponseLoading ? (
                                 <LoadingState />
                             ) : (
                                 listEmptyState
                             )}
-                        </div>
-                        <div className="shrink-0 relative flex justify-between items-center gap-0.5 whitespace-nowrap">
-                            {footerActions}
                         </div>
                     </div>
                     {featureFlags[FEATURE_FLAGS.MAX_SESSION_SUMMARIZATION_BUTTON] && (
@@ -211,7 +292,7 @@ export function Playlist({
                             fullWidth
                             size="small"
                             className="mt-2"
-                            disabledReason={!firstRecordingItem ? 'No recordings in the list' : undefined}
+                            disabledReason={!firstItem ? 'No recordings in the list' : undefined}
                         >
                             Summarize these recordings
                             <LemonTag type="warning" size="small" className="ml-auto uppercase">
@@ -252,6 +333,58 @@ const TitleWithCount = ({ title, count }: { title?: string; count: number }): JS
     )
 }
 
+const ListEmptyState = (): JSX.Element => {
+    const { sessionRecordingsAPIErrored, unusableEventsInFilter } = useValues(sessionRecordingsPlaylistLogic)
+
+    return (
+        <div className="p-3 text-sm text-secondary">
+            {sessionRecordingsAPIErrored ? (
+                <LemonBanner type="error">Error while trying to load recordings.</LemonBanner>
+            ) : unusableEventsInFilter.length ? (
+                <UnusableEventsWarning unusableEventsInFilter={unusableEventsInFilter} />
+            ) : (
+                <div className="flex flex-col gap-2">
+                    <SessionRecordingsPlaylistTroubleshooting />
+                </div>
+            )}
+        </div>
+    )
+}
+
+const CollectionEmptyState = ({
+    isSynthetic,
+    description,
+}: {
+    isSynthetic?: boolean
+    description?: string
+}): JSX.Element => {
+    const { sessionRecordingsAPIErrored, unusableEventsInFilter } = useValues(sessionRecordingsPlaylistLogic)
+
+    return (
+        <div className="p-3 text-sm text-secondary">
+            {sessionRecordingsAPIErrored ? (
+                <LemonBanner type="error">Error while trying to load recordings.</LemonBanner>
+            ) : unusableEventsInFilter.length ? (
+                <UnusableEventsWarning unusableEventsInFilter={unusableEventsInFilter} />
+            ) : isSynthetic ? (
+                <div className="flex flex-col gap-2">
+                    <h3 className="title text-secondary mb-0">No recordings yet</h3>
+                    <p>{description || 'This collection is automatically populated.'}</p>
+                </div>
+            ) : (
+                <div className="flex flex-col gap-2">
+                    <h3 className="title text-secondary mb-0">No recordings in this collection</h3>
+                    <p>
+                        To add recordings to this collection, go to the{' '}
+                        <Link to={urls.replay(ReplayTabs.Home)}>Recordings</Link> tab, click on a recording, then click
+                        "+ Add to collection" and select this collection from the list.
+                    </p>
+                </div>
+            )}
+        </div>
+    )
+}
+
 function SectionContent({
     section,
     loading,
@@ -263,7 +396,7 @@ function SectionContent({
     loading: boolean
     activeItemId: SessionRecordingType['id'] | null
     setActiveItemId: (item: SessionRecordingType) => void
-    emptyState: PlaylistProps['listEmptyState']
+    emptyState: JSX.Element
 }): JSX.Element {
     return 'content' in section ? (
         <>{section.content}</>
@@ -310,5 +443,32 @@ const LoadingState = (): JSX.Element => {
                 </div>
             ))}
         </>
+    )
+}
+
+/**
+ * TODO add docs on how to enrich custom events with session_id and link to it from here
+ */
+const UnusableEventsWarning = (props: { unusableEventsInFilter: string[] }): JSX.Element => {
+    return (
+        <LemonBanner type="warning">
+            <p>Cannot use these events to filter for session recordings:</p>
+            <li className="my-1">
+                {props.unusableEventsInFilter.map((event) => (
+                    <span key={event}>"{event}"</span>
+                ))}
+            </li>
+            <p>
+                Events have to have a <PropertyKeyInfo value="$session_id" /> to be used to filter recordings. This is
+                added automatically by{' '}
+                <Link to="https://posthog.com/docs/libraries/js" target="_blank">
+                    the Web SDK
+                </Link>
+                ,{' '}
+                <Link to="https://posthog.com/docs/libraries" target="_blank">
+                    and the Mobile SDKs (Android, iOS, React Native and Flutter)
+                </Link>
+            </p>
+        </LemonBanner>
     )
 }
